cursor doesn't erase rectangles anymore

+++ src/draw.c
@@ -87,7 +87,7 @@ void drawRect(int x_i, int y_i, int width, int height, uWord c){
 	}
 }
 
-void drawRectInBackdrop(int img[], int x_i, int y_i, int width, int height, uWord c){
+void drawRectInBackdrop(unsigned short img[], int x_i, int y_i, int width, int height, uWord c){
 	//struct fb_t* buffer = ((struct videoStruct*)VIDEO_BASE)->bfbp;
 	for(int y = y_i; y < height; ++y) {
 		for(int x = x_i; x < width; ++x) {
